Fixes for stupid long int size of 32bit in windows and other small fixes.

+++ hash/cpu/cpu_hasher.cpp
@@ -187,7 +187,7 @@ void cpu_hasher::__run() {
 
         if(!input.base.empty()) {
             hash_factory.set_seed_memory_offset(profile->memsize);
-            hash_factory.set_threads(argon2profile_default->memsize / profile->memsize);
+            hash_factory.set_threads((int)(argon2profile_default->memsize / profile->memsize));
 
             vector<string> hashes = hash_factory.generate_hashes(*profile, input.base, input.salt);
             for(vector<string>::iterator it = hashes.begin(); it != hashes.end(); ++it) {
